+++ src/main/java/com/getindata/connectors/http/internal/SinkHttpClientBuilder.java
@@ -1,11 +1,15 @@
 package com.getindata.connectors.http.internal;
 
 import java.io.Serializable;
+import java.util.Properties;
+
+import org.apache.flink.annotation.PublicEvolving;
 
 /**
  * Builder building {@link SinkHttpClient}.
  */
+@PublicEvolving
 public interface SinkHttpClientBuilder extends Serializable {
 
-    SinkHttpClient build();
+    SinkHttpClient build(Properties properties);
 }